move ui elements from genomics db to ui package
