+++ apps/genomicsdb/app/about/page.tsx
@@ -1,4 +1,3 @@
-import { MainLayout } from "@/components/main-layout";
 import { AboutPage } from "@/components/pages/about-page";
 
 export default function About() {
