+++ apps/genomicsdb/app/browse-datasets/page.tsx
@@ -1,4 +1,4 @@
-import { BrowseDatasetsPage } from "@/components/browse-datasets-page";
+import { BrowseDatasetsPage } from "@/components/pages/browse-datasets-page";
 
 export default function BrowseDatasets() {
     return <BrowseDatasetsPage />;
