+++ apps/genomicsdb/app/genome-browser/page.tsx
@@ -1,4 +1,3 @@
-import { MainLayout } from "@/components/main-layout";
 import { GenomeBrowserPage } from "@/components/pages/genome-browser-page";
 
 export default function GenomeBrowser() {
