+++ apps/genomicsdb/app/globals.css
@@ -389,3 +389,10 @@ a {
     background-color: var(--primary-blue);
     transition: width var(--transition-normal);
 }
+
+/* action buttons */
+.action-buttons {
+    display: flex;
+    gap: 0.75rem;
+    flex-wrap: wrap;
+}
