+++ apps/genomicsdb/app/layout.tsx
@@ -1,6 +1,6 @@
 import type React from "react";
 import type { Metadata } from "next";
-import { Inter, Roboto_Mono, Lato, Source_Sans_3 } from "next/font/google";
+import { Inter, Roboto_Mono, Lato } from "next/font/google";
 import { LoadingProvider } from "@/components/loading-context";
 import { ConditionalMainLayout } from "@/components/conditional-main-layout";
 import "./globals.css";
