+++ apps/genomicsdb/package.json
@@ -14,7 +14,8 @@
         "next": "15.3.3",
         "react": "^19.1.0",
         "react-dom": "^19.1.0",
-        "recharts": "^3.0.0"
+        "recharts": "^3.0.0",
+        "@niagads/ui": "^1.0.1-alpha.0"
     },
     "devDependencies": {
         "@eslint/eslintrc": "^3",
