+++ packages/UI/src/client/index.ts
@@ -1,6 +1,8 @@
 import "../styles/global.css";
+
 export * from "./Dropdown";
 export * from "./HelpIcon";
 export * from "./Tooltip";
 export * from "./Tabs";
 export * from "./Slider";
+export * from "./Footer";
