+++ packages/UI/src/index.ts
@@ -9,3 +9,6 @@ export * from "./RadioButton";
 export * from "./Skeleton";
 export * from "./Select";
 export * from "./TextInput";
+export * from "./loading-spinner";
+export * from "./SimpleTooltip";
+export * from "./ActionButton";
