+++ packages/UI/src/loading-spinner.tsx
@@ -1,9 +1,11 @@
+import React from "react";
+
 interface LoadingSpinnerProps {
     message?: string;
     size?: "sm" | "md" | "lg";
 }
 
-export function LoadingSpinner({ message = "Loading...", size = "md" }: LoadingSpinnerProps) {
+export const LoadingSpinner = ({ message = "Loading...", size = "md" }: LoadingSpinnerProps) => {
     const sizeClasses = {
         sm: "w-4 h-4",
         md: "w-8 h-8",
