+++ packages/UI/src/styles/action-button.css
@@ -1,9 +1,3 @@
-.action-buttons {
-    display: flex;
-    gap: 0.75rem;
-    flex-wrap: wrap;
-}
-
 .action-button {
     padding: 0.5rem 1rem;
     background-color: var(--surface);
