+++ packages/UI/src/styles/tooltip.css
